Organized the code for better readability. Added the month changing as you scroll through event dates
