+++ SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/ui/pages/home/Main View/DateHelper.swift
@@ -1,0 +1,50 @@
+//
+//  DateHelper.swift
+//  SHPE-UF-Mobile-Swift
+//
+//  Created by Matthew Segura on 3/22/24.
+//
+import SwiftUI
+
+struct DateHelper {
+    // Gets the current month as a string
+    func getCurrentMonth() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MMMM"
+        return dateFormatter.string(from: Date())
+    }
+    //Gets the month of an event as a string
+    func getMonth(for date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MMMM"
+        return dateFormatter.string(from: date)
+    }
+
+    // Gets the time in hour and minutes AM/PM format
+    func getTime(for date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "hh:mm a"
+        return dateFormatter.string(from: date)
+    }
+
+    // Gets the full date format for an event
+    func getDayFull(for date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MMM dd, yyyy"
+        return dateFormatter.string(from: date)
+    }
+
+    // Gets the day abbreviation (e.g., Mon, Tue)
+    func getDayAbbreviation(for date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "E"
+        return dateFormatter.string(from: date)
+    }
+
+    // Gets the day number from a date
+    func getDayNumber(for date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "d"
+        return dateFormatter.string(from: date)
+    }
+}
