+++ SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/ui/pages/home/Main View/HomePageContentView.swift
@@ -19,7 +19,6 @@ struct HomePageContentView: View {
                 .tag(1)
                 .tabItem {
                     Image(selectedTab == 1 ? "clicked_leaderboard" : colorScheme == .dark ? "dark_leaderboard":"Leaderboard")
-
                 }
             
             Text("Profile")
